requirements.txt
beautifulsoup4==4.14.2
chromadb==1.1.1
groq==0.32.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.78
langchain-groq==0.3.8
langchain-text-splitters==0.3.11
numpy==2.3.3
pandas==2.3.3
python-dotenv==1.1.1
requests==2.32.5
